冷蔵庫クラスと、冷蔵室クラスの追加。ただし、今のままだと食品クラスが使われていない...？

+++ src/main/java/com/example/demo/model/Shokuhin.java
@@ -7,7 +7,7 @@ public class Shokuhin {
 	private int id;
     private String name;
     private String genre;
-    private double amount;     // ← 数量（数値）
+    private int amount;     // ← 数量（数値）
     private String unit;       // ← 単位（個、本、gなど）
     private String expiryDate;
 
@@ -15,7 +15,7 @@ public class Shokuhin {
     	this.id = idCounter++;
     }
 
-    public Shokuhin(String name, String genre, double amount, String unit, String expiryDate) {
+    public Shokuhin(String name, String genre, int amount, String unit, String expiryDate) {
         this.id = idCounter++;
         this.name = name;
         this.genre = genre;
@@ -31,13 +31,13 @@ public class Shokuhin {
     // getter & setter
     public String getName() { return name; }
     public String getGenre() { return genre; }
-    public double getAmount() { return amount; }
+    public int getAmount() { return amount; }
     public String getUnit() { return unit; }
     public String getExpiryDate() { return expiryDate; }
 
     public void setName(String name) { this.name = name; }
     public void setGenre(String genre) { this.genre = genre; }
-    public void setAmount(double amount) { this.amount = amount; }
+    public void setAmount(int amount) { this.amount = amount; }
     public void setUnit(String unit) { this.unit = unit; }
     public void setExpiryDate(String expiryDate) { this.expiryDate = expiryDate; }
